+++ src/Entity/Season.php
@@ -121,7 +121,7 @@ class Season
         return $this->isActive;
     }
 
-    public function setActive(bool $isActive): static
+    public function setIsActive(bool $isActive): static
     {
         $this->isActive = $isActive;